Moved Math files into the separate folder
